♻️ move mockable to @datadog/js-core/util
